backend/requirements.txt
bson==0.5.8
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
gcloud==0.17.0
googleapis-common-protos==1.6.0
httplib2==0.13.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jws==0.1.3
MarkupSafe==1.1.1
oauth2client==3.0.0
protobuf==3.9.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dateutil==2.8.0
python-dotenv==0.10.3
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.12.0
smartcar==4.1.1
urllib3==1.25.3
Werkzeug==0.15.5
